+++ Sources/SwiftGenKit/Parsers/Strings/StringsParser.swift
@@ -67,7 +67,8 @@ public enum Strings {
     /// This list must be ordered by priority, higher priority first
     private static let subParsers: [StringsFileTypeParser.Type] = [
       StringsDictFileParser.self,
-      StringsFileParser.self
+      StringsFileParser.self,
+      StringsCatalogFileParser.self
     ]
 
     public init(options: [String: Any] = [:], warningHandler: Parser.MessageHandler? = nil) throws {